Automate ETag, If-None-Match, If-Match handling in response processing.
Restructure response processing so that small enough responses are
buffered fully in memory up to a configurable threshold. Compute ETag
and automatically handle If-Match, If-None-Match for GET and HEAD
requests below the threshold size. For larger responses automatically
switch back to chunked encoding, and suppress If-Match, If-None-Match
handling but still compute ETag and output it in trailer headers.

From: Lassi Tuura <lat@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/SiteDB/trunk@14802 4525493e-7705-40b1-a816-d608a930855b