Pick expire time limit defaults from cherrypy expire tool.
This is needed because cherrypy never actually runs the REST entity
method as a URL handler, so it never executes any of the hooks
attached by tools.  For now simulate the behaviour of the 'expires'
tool only, and even then in part; probably need to introduce a more
general way to execute all normal hook transitions for entities.

From: Lassi Tuura <lat@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/SiteDB/trunk@14806 4525493e-7705-40b1-a816-d608a930855b